moved _stream function for SSE messages to base client
